Structs are now at the top-level Concurrent namespace.
