+++ lib/concurrent.rb
@@ -9,7 +9,6 @@ require 'concurrent/collections'
 require 'concurrent/errors'
 require 'concurrent/executors'
 require 'concurrent/utilities'
-require 'concurrent/struct'
 
 require 'concurrent/atomic/atomic_reference'
 require 'concurrent/atom'
@@ -17,11 +16,14 @@ require 'concurrent/async'
 require 'concurrent/dataflow'
 require 'concurrent/delay'
 require 'concurrent/future'
+require 'concurrent/immutable_struct'
 require 'concurrent/ivar'
 require 'concurrent/maybe'
+require 'concurrent/mutable_struct'
 require 'concurrent/mvar'
 require 'concurrent/promise'
 require 'concurrent/scheduled_task'
+require 'concurrent/settable_struct'
 require 'concurrent/timer_task'
 require 'concurrent/tvar'
