--- lib/concurrent/struct.rb
@@ -1,3 +1,0 @@
-require 'concurrent/struct/immutable_struct'
-require 'concurrent/struct/mutable_struct'
-require 'concurrent/struct/settable_struct'
