+++ spec/concurrent/immutable_struct_spec.rb
