+++ spec/concurrent/struct_shared.rb
